requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
allennlp==0.9.0
apptools==5.1.0
argh==0.26.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn==21.2.1
autograd==1.3
Automat==20.2.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.252
botocore==1.12.252
boxes==0.0.1
brotlipy==0.7.0
cachetools==3.1.1
certifi==2021.5.30
Click==7.0
configobj==5.0.6
configparser==4.0.2
conllu==1.3.1
constantly==15.1.0
cycler==0.10.0
cymem==2.0.2
d2l==0.17.0
dataclasses==0.8
datasets==0.0.1
defusedxml==0.6.0
dgl==0.6.1
docker-pycreds==0.4.0
docutils==0.15.2
editdistance==0.5.3
entrypoints==0.3
envisage==6.0.1
filelock==3.0.12
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
future==0.18.2
gdown==3.8.3
gensim==3.8.1
geomstats==2.2.2
geoopt==0.3.1
gevent==1.4.0
giotto-ph==0.1.0
giotto-tda==0.5.1
git-lfs==1.6
gitdb2==2.0.6
GitPython==3.0.5
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
gql==0.1.0
graphql-core==2.2.1
greenlet==0.4.15
h5py==2.10.0
httplib2==0.14.0
huggingface-hub==0.0.16
hyperlink==21.0.0
idna-ssl==1.1.0
imagesize==1.1.0
importlib-metadata==0.23
importlib-resources==5.2.2
incremental==21.3.0
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==1.0.1
jsonnet==0.14.0
jsonpickle==1.2
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
logger==1.4
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==7.2.0
mpmath==1.2.1
multidict==5.1.0
murmurhash==1.0.2
mypy==0.701
mypy-extensions==0.4.1
nbconvert==5.6.0
nbformat==4.4.0
ndim==0.1.5
nltk==3.4.5
notebook==6.0.1
numpydoc==0.9.1
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
orthopy==0.9.3
overrides==2.0
packaging==21.0
pandas==1.0.3
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
plac==0.9.6
plotly==5.3.1
pluggy==0.13.0
preshed==2.0.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.17.3
psutil==5.6.5
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pydot==1.4.2
pyface==7.3.0
pyflagser==0.4.4
Pygments==2.9.0
pymanopt==0.2.5
pyNTCIREVAL==0.0.3
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==5.4.3
python-dateutil==2.8.1
python-igraph==0.9.6
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
quadpy==0.16.9
regex==2021.7.6
responses==0.13.3
retrying==1.3.3
rsa==4.7.2
Rx==1.6.1
s3transfer==0.2.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.7.1
sentencepiece==0.1.96
sentry-sdk==1.3.0
shortuuid==1.0.1
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
smmap2==3.0.1
snowballstemmer==2.1.0
spacy==2.1.9
Sphinx==4.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
srsly==1.0.5
subprocess32==3.5.4
sympy==1.8
tenacity==8.0.1
tensorboardX==2.4
terminado==0.10.1
testpath==0.5.0
textblob==0.15.3
texttable==1.6.4
thinc==7.0.8
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.8.1+cu111
torchaudio==0.8.1
torchtext==0.10.0
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
traits==6.2.0
traitsui==7.2.1
transformers==4.10.2
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.3.5
typing-extensions==3.10.0.0
Unidecode==1.2.0
uritemplate==3.0.1
vtk==9.0.3
Wand==0.6.7
wandb==0.12.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
word2number==1.1
wslink==1.0.7
yarl==1.6.3
zipp==3.5.0
zope.interface==5.4.0
